int p
p being size_t was not really necessary,
and it makes it easy to confuse with things
that really are byte counts like k.

+++ gfx.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <stddef.h>
-
 #if defined(__ARM_FEATURE_FP16_VECTOR_ARITHMETIC)
     #define N 8
     typedef _Float16 half;
@@ -31,7 +29,7 @@ typedef struct {
     F32  x,y;
 } Cold;
 
-typedef RGBA (ABI Effect)(void* ctx, size_t p, RGBA src, Cold*);
+typedef RGBA (ABI Effect)(void* ctx, int p, RGBA src, Cold*);
 
 Effect
     blend_dst,
